backend/requirements.txt
aiosqlite==0.20.0
fastapi==0.111.0
pwdlib[argon2]==0.2.0
pydantic==2.7.4
PyJWT==2.8.0
sqlalchemy[asyncio]==2.0.30
uvicorn==0.30.1
pyhumps==3.8.0